PROTON-1454: C proactor API doc overhaul, minor API improvements.
Reviewed and improved all API doc in proactor.h and listener.h

API improvements:
- Reverted a85c89a "PROTON-1445: Change proactor ownership model" - it was a bad idea.
- pn_proactor_connect|listen return void, all errors reported by events
- rename pn_proactor_disown_connection -> pn_proactor_release_connection
- document the list of event types used by the proactor.
- const fixes for pn_proactor_addr functions
- verify all proactor library functions are in proactor.h or listener.h

+++ proton-c/src/tests/test_tools.h
@@ -132,7 +132,7 @@ bool test_etype_equal_(test_t *t, pn_event_type_t want, pn_event_type_t got, con
 
 void print_bad_etypes(const char* prefix, const pn_event_type_t* seq, size_t len, size_t bad) {
   fprintf(stderr, "%s", prefix);
-  for (int i = 0; i < len; ++i) {
+  for (size_t i = 0; i < len; ++i) {
     fprintf(stderr, (i == bad) ? ">>>>%s" : "%s", pn_event_type_name(seq[i]));
     if (i < len-1) fprintf(stderr, ", ");
   }
@@ -142,7 +142,7 @@ void print_bad_etypes(const char* prefix, const pn_event_type_t* seq, size_t len
 
 bool test_etypes_equal_(test_t *t, const pn_event_type_t* want, size_t want_len, const pn_event_type_t* got, size_t got_len, const char *file, int line) {
   size_t len = want_len < got_len ? want_len : got_len;
-  for (int i = 0; i < len; ++i) {
+  for (size_t i = 0; i < len; ++i) {
     if (want[i] != got[i]) {
       test_errorf_(t, NULL, file, line, "event sequences don't match:");
       print_bad_etypes("  want: ", want, want_len, i);
@@ -261,7 +261,7 @@ int sock_port(sock_t sock) {
   return ntohs(port);
 }
 
-/* Combines includes a sock_t with the int and char* versions of the port for convenience */
+/* Combines a sock_t with the int and char* versions of the port for convenience */
 typedef struct test_port_t {
   sock_t sock;
   int port;                     /* port as integer */
